Create FlagImage view
Project 3 sets a challenge to add a FlagImage view to the GuessTheFlag project. This change refactors the original content view to extract the Image and its modifiers that display the flag. This change also moves view code from the main ContentView body into three computed properties to make the code easier to follow.