Project 6 challenge 1
The challenge was to modify the GuessTheFlag app so that clicking the correct flag resulted in it spinning on its y-axis. I got that working but it looked poor if the alert covered the animation if the middle flag was the correct answer.

More state was added to distinguish between a correct and an incorrect guess so that the alert only appeared for the incorrect case. So far, so good. Then came the problem of resetting the game state for the next round without instantly zapping the animation. This led to the use of DispatchQueue.main.asyncAfter to execute the game reset function after a delay that would allow the animation to run. The final step was to disable buttons while the animation was running otherwise it was possible to change a correct guess to an incorrect one and that just looked weird.