Project 6 challenge 3
The final challenge from project 6 was to animate an incorrect selection in a manner of our own choosing. I chose to spin the flag counter-clockwise about its z-axis while scaling it to 0. I did a bit of re-working to scale a correct selection to 1.25 while setting the opacity of the other flags to 0.25. I also tidied up the code a bit (at least in my eyes).